+++ lib/src/widget/widget.dart
@@ -1,0 +1,1 @@
+export 'result_table/result_table.dart';